requirements.txt
aniso8601==9.0.1
apispec==5.1.1
cffi==1.15.0
click==8.0.4
cryptography==36.0.2
dateutils==0.6.12
dnspython==2.2.1
email-validator==1.1.3
Flask==2.0.3
flask-apispec==0.11.0
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-WTF==1.0.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
marshmallow==3.15.0
mongoengine==0.24.0
packaging==21.3
pycparser==2.21
pymongo==4.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
webargs==8.1.0
Werkzeug==2.0.3
WTForms==3.0.1
